refactor: split socket handlers by feature (#4894)
